blind-to-x/requirements.txt
# Pinned for CI stability — update via Dependabot PRs
playwright==1.50.0
anthropic==0.43.0
notion-client==2.2.1
PyYAML==6.0.2
requests==2.32.3
python-dotenv==1.0.1
cloudinary==1.41.0
curl_cffi==0.14.1
beautifulsoup4==4.12.3
openai==1.59.9
tweepy==4.14.0
aiohttp==3.11.11
joblib>=1.3.0
pytest-asyncio==0.24.0
